UIBarButtonItem removeTitleShadow now also removes the shadow in control state highlighted

+++ Classes/ios/UIImage+FlatUI.m
@@ -138,7 +138,7 @@ static CGFloat edgeSizeFromCornerRadius(CGFloat cornerRadius) {
     ctrlPoint.y += radius;
     mPoint.x += radius;
     mPoint.y += radius;
-    if (radius > 0) [path addArcWithCenter:ctrlPoint radius:radius startAngle:M_PI + M_PI_2 endAngle:0 clockwise:YES];
+    if (radius > 0) [path addArcWithCenter:ctrlPoint radius:radius startAngle:(float)M_PI + (float)M_PI_2 endAngle:0 clockwise:YES];
     
     mPoint.y = CGRectGetMaxY(rect) - radius;
     [path addLineToPoint:mPoint];
@@ -147,7 +147,7 @@ static CGFloat edgeSizeFromCornerRadius(CGFloat cornerRadius) {
     mPoint.y += radius;
     mPoint.x -= radius;
     ctrlPoint.x -= radius;
-    if (radius > 0) [path addArcWithCenter:ctrlPoint radius:radius startAngle:0 endAngle:M_PI_2 clockwise:YES];
+    if (radius > 0) [path addArcWithCenter:ctrlPoint radius:radius startAngle:0 endAngle:(float)M_PI_2 clockwise:YES];
     
     mPoint.x = rect.origin.x + (10.0f);
     [path addLineToPoint:mPoint];
